[FIX] Changing Memory from int to long

+++ hyracks-fullstack/hyracks/hyracks-dataflow-std/src/main/java/org/apache/hyracks/dataflow/std/buffermanager/IFramePool.java
@@ -27,7 +27,7 @@ public interface IFramePool {
 
     int getMinFrameSize();
 
-    int getMemoryBudgetBytes();
+    long getMemoryBudgetBytes();
 
     /**
      * Get a frame of given size. <br>
